[style] Refactor naming
